backend/kafka_handlers/detect_consumer/requirements.txt
opencv-python==4.11.0.86
ultralytics==8.3.111
kafka-python==2.2.10
dotenv==0.9.9
minio==7.2.15
